refactor: credential status utils and selector
