+++ ts/features/itwallet/credentials/store/selectors/index.ts
@@ -1,17 +1,17 @@
+import { pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/lib/Option";
 import { createSelector } from "reselect";
-import { pipe } from "fp-ts/lib/function";
-import { Errors } from "@pagopa/io-react-native-wallet";
 import { GlobalState } from "../../../../../store/reducers/types";
+import { getFiscalCodeFromCredential } from "../../../common/utils/itwClaimsUtils";
+import {
+  getCredentialStatus,
+  getCredentialStatusObject
+} from "../../../common/utils/itwCredentialStatusUtils";
 import { CredentialType } from "../../../common/utils/itwMocksUtils";
 import {
   ItwJwtCredentialStatus,
   StoredCredential
 } from "../../../common/utils/itwTypesUtils";
-import {
-  getCredentialStatus,
-  getFiscalCodeFromCredential
-} from "../../../common/utils/itwClaimsUtils";
 
 export const itwCredentialsSelector = (state: GlobalState) =>
   state.features.itWallet.credentials;
@@ -85,23 +85,7 @@ export const itwCredentialStatusSelector = createSelector(
       return { status: undefined, message: undefined };
     }
 
-    const { storedStatusAttestation, issuerConf, credentialType } =
-      credentialOption.value;
-
-    const errorCode =
-      storedStatusAttestation?.credentialStatus === "invalid"
-        ? storedStatusAttestation.errorCode
-        : undefined;
-
-    return {
-      status: getCredentialStatus(credentialOption.value),
-      message: errorCode
-        ? Errors.extractErrorMessageFromIssuerConf(errorCode, {
-            issuerConf,
-            credentialType
-          })
-        : undefined
-    };
+    return getCredentialStatusObject(credentialOption.value);
   }
 );
